requirements.txt
datasets==4.4.1
faiss_gpu==1.7.2
fastapi==0.121.1
numpy==2.3.4
pandas==2.3.3
peft==0.17.1
pydantic==2.12.4
Requests==2.32.5
scikit_learn==1.7.2
sentence_transformers==5.1.2
streamlit==1.51.0
torch==2.5.1+cu121
tqdm==4.67.1
transformers==4.57.1
trl==0.25.0
uvicorn==0.38.0
